requirements.txt
# Core agent framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# HTTP client
requests>=2.31.0

# Database
mysql-connector-python>=9.0.0

# Environment variable management
python-dotenv>=1.0.0
# LLM provider
openai>=1.0.0

# Data validation
pydantic>=2.0.0